requirements.txt
# HEADY_BRAND:BEGIN
# HEADY SYSTEMS :: SACRED GEOMETRY
# FILE: requirements.txt
# LAYER: root
# 
#         _   _  _____    _  __   __
#        | | | || ____|  / \ \  / /
#        | |_| ||  _|   / _ \ \ V / 
#        |  _  || |___ / ___ \ | |  
#        |_| |_||_____/_/   \_\|_|  
# 
#    Sacred Geometry :: Organic Systems :: Breathing Interfaces
# HEADY_BRAND:END

psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.31.0
numpy>=1.24.0
transformers>=4.30.0
torch>=2.0.0
flask>=2.3.0
asyncio>=3.4.3
aiohttp>=3.8.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
